Issue #68 refactor: notification date 계산 확장 함수 추가
